ui_detection_multimodel/requirements.txt
ultralytics
torch
opencv-python
pyyaml
matplotlib